% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanced_ternary.R
\name{balanced_ternary}
\alias{balanced_ternary}
\title{Balanced Ternary}
\usage{
balanced_ternary(n)
}
\arguments{
\item{n}{The number of terms of the sequence to generate.}
}
\description{
Write a number n in base 3, substitute every 2 with a -1, then convert to
decimal.
}
\examples{
bt <- balanced_ternary(200)
plot(bt$n, bt$y)

}
\references{
This sequence in the On-Line Encyclopedia of Integer Sequences:
  \href{https://oeis.org/A117966}{A117966}.

  The Numberphile video featuring Neil Sloane and Brady Haran.
  \href{https://www.youtube.com/watch?v=o8c4uYnnNnc}{Amazing Graphs II
  (including Star Wars)}.
}
