% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chaos_game.R
\name{chaos_fern}
\alias{chaos_fern}
\title{Chaos Fern}
\usage{
chaos_fern(
  params = c("barnsley", "thelypteridaceae", "leptosporangiate"),
  iterations = 10000
)
}
\arguments{
\item{params}{Either the name of a builtin parameter set (see details) or a
data fram with the columns a, b, c, d, e and p.}

\item{iterations}{The number of iterations.}
}
\description{
Create a fractal that resembles a fern.
}
\details{
The builtin parameter sets are "barnsley", "thelypteridaceae" and
"leptosporangiate". The names can be abbreviated. The barnsley fern is the
one shown in the video. The two others are taken from the examples on
\href{https://en.wikipedia.org/wiki/Barnsley_fern}{wikipedia}.
}
\examples{
ferns <- list(
    chaos_fern(iterations = 30000),
    chaos_fern("thelypteridaceae", iterations = 30000),
    chaos_fern("leptosporangiate", iterations = 30000)
)
par(mfrow = c(1, 3), mar = rep(0, 4))
for (fern in ferns) {
    plot(fern, pch = ".", axes = FALSE, ann = FALSE, asp = 1, col = "forestgreen")
}

}
\references{
The Numberphile video featuring Ben Sparks and Brady Haran:
  \href{https://youtu.be/kbKtFN71Lfs?t=338}{Chaos Game} (skip to 5:38).
}
