% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chaos_game.R
\name{chaos_polygon}
\alias{chaos_polygon}
\title{Chaos Polygon}
\usage{
chaos_polygon(n_vertices = 3, proportion = 0.5, n_iterations = 10000)
}
\arguments{
\item{n_vertices}{The number of vertices.}

\item{proportion}{The proportion of the distance to move from the point to
the target vertex.}

\item{n_iterations}{The number of iterations.}
}
\description{
Start with the vertices of a regular polygon. Randomly place a tracepoint at
one of them. Each step of the iteration choose a random vertex and move the
tracepoint some distance towards that.
}
\examples{
p <- list(
    chaos_polygon(),
    chaos_polygon(4, 0.55),
    chaos_polygon(5, 0.62),
    chaos_polygon(6, 2/3),
    chaos_polygon(30, 0.91),
    chaos_polygon(proportion = 1.5)
)
par(mfrow = c(2, 3), mar = rep(0, 4))
for (xy in p) {
    plot(xy, pch = 20, axes = FALSE, ann = FALSE, asp = 1, cex = 0.25)
}

}
\references{
The Numberphile video featuring Ben Sparks and Brady Haran:
  \href{https://youtu.be/kbKtFN71Lfs}{Chaos Game}.
}
