% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fly_straight.R
\name{fly_straight}
\alias{fly_straight}
\title{Fly Straight}
\usage{
fly_straight(n)
}
\arguments{
\item{n}{The number of terms of the sequence to generate.}
}
\description{
TODO: write a nice description
}
\examples{
fs <- fly_straight(1000)
plot(fs)

}
\references{
This sequence in the On-Line Encyclopedia of Integer Sequences:
  \href{https://oeis.org/A133058}{A133058}.

  The Numberphile video featuring Neil Sloane and Brady Haran:
  \href{https://www.youtube.com/watch?v=pAMgUB51XZA}{Amazing Graphs}.
}
