% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_fire.R
\name{forest_fire}
\alias{forest_fire}
\title{Forest Fire}
\usage{
forest_fire(n)
}
\arguments{
\item{n}{The number of terms of the sequence to generate.}
}
\description{
Sequence of positive integers where each is chosen to be as small as possible
subject to the condition that no three terms a(j), a(j+k), a(j+2k) (for any j
and k) form an arithmetic progression.
}
\examples{
ff <- forest_fire(5000)
plot(ff, pch = 16, col = "#00000070")

}
\references{
This sequence in the On-Line Encyclopedia of Integer Sequences:
  \href{https://oeis.org/A229037}{A229037}.

  The Numberphile video featuring Neil Sloane and Brady Haran.
  \href{https://youtu.be/o8c4uYnnNnc?t=217}{Amazing Graphs II
  (including Star Wars)} (skip to 3:37).
}
