% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcd.R
\name{gcd}
\alias{gcd}
\title{Greatest Common Divisor}
\usage{
gcd(a, b)
}
\arguments{
\item{a, b}{Single integers.}
}
\value{
The greatest common divisor of \code{a} and \code{b}.
}
\description{
Compute the greatest common divisor.
}
\examples{
gcd(12, 8)

}
