% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcd.R
\name{get_gcd}
\alias{get_gcd}
\title{Greates Common Divisor}
\usage{
get_gcd(a, b)
}
\arguments{
\item{a, b}{Single numbers.}
}
\description{
Recursively compute the greatest common divisor.
}
\examples{
get_gcd(12, 6)

}
