% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/look_and_say.R
\name{look_and_say}
\alias{look_and_say}
\title{Look-and-Say Sequence}
\usage{
look_and_say(n = 1, seed = 1, return_all = FALSE)
}
\arguments{
\item{n}{The number of iteration steps.}

\item{seed}{The starting sequence.}

\item{return_all}{Logical. Should all steps be returned? Defaults to
\code{TRUE}.}
}
\value{
The final iteration of the sequence as a vector if \code{return_all}
  is \code{FALSE}, otherwise a list of all steps of the sequence.
}
\description{
A sequence which describes its previous term.
}
\examples{
look_and_say(seed = 1, n = 7, return_all = TRUE)

}
\references{
This sequence in the On-Line Encyclopedia of Integer Sequences:
  \href{https://oeis.org/A005150}{A005150}.

  The Numberphile video featuring John Conway and Brady Haran:
  \href{https://youtu.be/ea7lJkEhytA}{Look-and-Say Numbers (feat John
  Conway)}.
}
