% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucky_number_sieve.R
\name{lucky_number_sieve}
\alias{lucky_number_sieve}
\title{Lucky Number Sieve}
\usage{
lucky_number_sieve(n)
}
\arguments{
\item{n}{The upper limit of the region that is searched for lucky numbers.}
}
\value{
A vector of the lucky numbers between 1 and n.
}
\description{
Get all lucky numbers <= n.
}
\examples{
lucky_number_sieve(100)

}
\references{
This sequence in the On-Line Encyclopedia of Integer Sequences:
  \href{https://oeis.org/A000959}{A000959}.

  The Numberphile video featuring Ria Symonds and Brady Haran:
  \href{https://youtu.be/RxxDD2LWAyY}{What is a lucky number?}.
}
