% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Numberphile-package.R
\docType{package}
\name{Numberphile-package}
\alias{Numberphile}
\alias{Numberphile-package}
\title{Numberphile: A Package Inspired by Numberphile Videos}
\description{
Functions to recreate some neat sequences and
    diagrams seen in Numberphile videos on youtube.
}
\author{
\strong{Maintainer}: Sebastian Henz \email{bastihz.dev@posteo.de}

}
