% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recaman_sequence.R
\name{recaman}
\alias{recaman}
\title{Recaman Sequence}
\usage{
recaman(n, plot = FALSE)
}
\arguments{
\item{n}{The number of terms of the sequence to generate.}

\item{plot}{Logical. Should the sequence be plotted with semicircles?}
}
\value{
The Recamán sequence.
}
\description{
Always go back. If you can't go back, go forward.
}
\examples{
r <- recaman(100, TRUE)

}
\references{
This sequence in the On-Line Encyclopedia of Integer Sequences:
  \href{https://oeis.org/A005132}{A005132}.

  The Numberphile video featuring Alex Bellos and Brady Haran:
  \href{https://youtu.be/FGC5TdIiT9U}{The Slightly Spooky Recaman Sequence}.
}
