% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sieve_of_eratosthenes.R
\name{sieve_of_eratosthenes}
\alias{sieve_of_eratosthenes}
\title{Sieve of Eratosthenes}
\usage{
sieve_of_eratosthenes(p_max)
}
\arguments{
\item{p_max}{The upper limit of the region that is searched for primes.}
}
\value{
A vector of the primes between 2 and n.
}
\description{
Get all prime numbers between 2 and n.
}
\examples{
sieve_of_eratosthenes(100)

}
