% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_sum_problem.R
\name{square_sum_problem}
\alias{square_sum_problem}
\title{Square Sum Problem}
\usage{
square_sum_problem(n)
}
\arguments{
\item{n}{The length of the chain.}
}
\description{
Using the integers from 1 to n construct a chain such that each pair of
neighboring numbers adds up to a square number.
}
\details{
The function returns the first valid chain that is found. Maybe some day I
will write a version that returns all possible chains.
}
\examples{
square_sum_problem(15)

}
\references{
The Numberphile video featuring Matt Parker and Brady Haran:
  \href{https://youtu.be/G1m7goLCJDY}{The Square-Sum Problem}.
}
