% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulam_spiral.R
\name{ulam_spiral}
\alias{ulam_spiral}
\title{Ulam Spiral}
\usage{
ulam_spiral(sidelength)
}
\arguments{
\item{sidelength}{The side length of the spiral.}
}
\description{
Write the numbers from 1 to n in a square spiral and mark all primes.
}
\examples{
us <- ulam_spiral(100)
grid::grid.raster(!us, interpolate = FALSE)

}
\references{
The Numberphile video featuring James Grime and Brady Haran:
  \href{https://youtu.be/iFuR97YcSLM}{Prime Spirals}.
}
