% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberphile-package.R
\docType{package}
\name{numberphile-package}
\alias{numberphile}
\alias{numberphile-package}
\title{numberphile: Inspired by Numberphile Videos}
\description{
Create some neat sequences and diagrams seen in videos on the 
    Numberphile youtube channel.
}
\author{
\strong{Maintainer}: Sebastian Henz \email{bastihz.dev@posteo.de}

}
