% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime_parallelograms.R
\name{prime_parallelograms}
\alias{prime_parallelograms}
\title{Prime Parallelograms}
\usage{
prime_parallelograms(p_max, plot = FALSE)
}
\arguments{
\item{p_max}{The upper limit of the region that is searched for primes.}

\item{plot}{Logical. Should the parallelograms be plotted?}
}
\value{
A data frame with columns for the primes, the result of the binary
  reversal and substraction, and the indices of the primes.
}
\description{
Take a prime number, convert it to binary, reverse it, convert it back to
decimal, and then substract it from the original prime number.
}
\examples{
pp <- prime_parallelograms(10000)
plot(pp$i, pp$y, pch = 20, cex = 0.3, las = 1)

}
\references{
This sequence in the On-Line Encyclopedia of Integer Sequences:
  \href{https://oeis.org/A265326}{A265326}.

  The Numberphile video featuring Neil Sloane and Brady Haran.
  \href{https://youtu.be/pAMgUB51XZA?t=467}{Amazing Graphs} (skip to 7:47).
}
